+++ tests/test_arguments.py
@@ -1,5 +1,3 @@
-# -*- encoding: utf-8 -*-
-# Version 2.2.a
 
 import os
 import sys
@@ -19,12 +17,14 @@ def some_params() -> args.HighlighterParams:
         target_fasta_fpath=None,
         bam_fpath=None,
         outfpath=os.path.join(os.getcwd(), 'highlighted_sequence.gbk'),
-        coverage_thresholds=(10, 15, 20),
+        lower_coverage_thresholds=(10, 15, 20),
+        upper_coverage_coefficients=(1.5, 2.0),
         suppress_zero_cov_output=False,
+        min_feature_len=5,
         topology='linear',
         organism='.'
     )
-# end def some_params
+# end def
 
 
 # === Fixtures for function `src.arguments._parse_options` ===
@@ -36,9 +36,10 @@ def opts_all_is_ok_short_options(test_fasta_fpath, test_bam_fpath, test_outfpath
         ['-b', test_bam_fpath],
         ['-o', test_outfpath],
         ['-c', '10,20'],
+        ['-C', '1.5,2.0'],
         ['-n', ''],
     ]
-# end def opts_all_is_ok_short_options
+# end def
 
 @pytest.fixture
 def opts_all_is_ok_long_options(test_fasta_fpath, test_bam_fpath, test_outfpath) -> List[List[str]]:
@@ -46,12 +47,13 @@ def opts_all_is_ok_long_options(test_fasta_fpath, test_bam_fpath, test_outfpath)
         ['--target-fasta', test_fasta_fpath],
         ['--bam', test_bam_fpath],
         ['--outfile', test_outfpath],
-        ['--coverage-thresholds', '10,20'],
+        ['--lower-coverage-thresholds', '10,20'],
+        ['--upper-coverage-coefficients', '1.5,2.0'],
         ['--no-zero-output', ''],
         ['--circular', ''],
         ['--organism', 'Czort lysy']
     ]
-# end def opts_all_is_ok_long_options
+# end def
 
 @pytest.fixture
 def opts_invalid_fasta(test_fasta_fpath, test_bam_fpath) -> List[List[str]]:
@@ -60,7 +62,7 @@ def opts_invalid_fasta(test_fasta_fpath, test_bam_fpath) -> List[List[str]]:
         ['-b', test_bam_fpath],
         ['-c', '10,20'],
     ]
-# end def opts_invalid_fasta
+# end def
 
 @pytest.fixture
 def opts_fasta_not_fasta(test_bam_fpath) -> List[List[str]]:
@@ -69,7 +71,7 @@ def opts_fasta_not_fasta(test_bam_fpath) -> List[List[str]]:
         ['-b', test_bam_fpath],
         ['-c', '10,20'],
     ]
-# end def opts_fasta_not_fasta
+# end def
 
 @pytest.fixture
 def opts_invalid_bam(test_fasta_fpath, test_bam_fpath) -> List[List[str]]:
@@ -78,7 +80,7 @@ def opts_invalid_bam(test_fasta_fpath, test_bam_fpath) -> List[List[str]]:
         ['-b', 'TRASH' + test_bam_fpath], # <- non-extant file
         ['-c', '10,20'],
     ]
-# end def opts_invalid_bam
+# end def
 
 @pytest.fixture
 def opts_bam_not_bam(test_fasta_fpath) -> List[List[str]]:
@@ -87,7 +89,7 @@ def opts_bam_not_bam(test_fasta_fpath) -> List[List[str]]:
         ['-b', test_fasta_fpath],
         ['-c', '10,20'],
     ]
-# end def opts_bam_not_bam
+# end def
 
 @pytest.fixture
 def opts_NAN_threshold(test_fasta_fpath, test_bam_fpath) -> List[List[str]]:
@@ -96,7 +98,7 @@ def opts_NAN_threshold(test_fasta_fpath, test_bam_fpath) -> List[List[str]]:
         ['-b', test_bam_fpath],
         ['-c', '10,2X'], # <- non-numeric threshold
     ]
-# end def opts_NAN_threshold
+# end def
 
 @pytest.fixture
 def opts_zero_threshold(test_fasta_fpath, test_bam_fpath) -> List[List[str]]:
@@ -105,7 +107,7 @@ def opts_zero_threshold(test_fasta_fpath, test_bam_fpath) -> List[List[str]]:
         ['-b', test_bam_fpath],
         ['-c', '0,20'], # <- zero threshold
     ]
-# end def opts_zero_threshold
+# end def
 
 @pytest.fixture
 def opts_negative_threshold(test_fasta_fpath, test_bam_fpath) -> List[List[str]]:
@@ -114,7 +116,7 @@ def opts_negative_threshold(test_fasta_fpath, test_bam_fpath) -> List[List[str]]
         ['-b', test_bam_fpath],
         ['-c', '-1,20'], # <- negative threshold
     ]
-# end def opts_negative_threshold
+# end def
 
 @pytest.fixture
 def opts_keep_zero_thresahold(test_fasta_fpath, test_bam_fpath) -> List[List[str]]:
@@ -123,7 +125,7 @@ def opts_keep_zero_thresahold(test_fasta_fpath, test_bam_fpath) -> List[List[str
         ['-b', test_bam_fpath],
         ['-c', '10,20'],
     ]
-# end def opts_keep_zero_thresahold
+# end def
 
 @pytest.fixture
 def opts_suppress_zero_thresahold(test_fasta_fpath, test_bam_fpath) -> List[List[str]]:
@@ -133,7 +135,7 @@ def opts_suppress_zero_thresahold(test_fasta_fpath, test_bam_fpath) -> List[List
         ['-c', '10,20'],
         ['-n', '']
     ]
-# end def opts_suppress_zero_thresahold
+# end def
 
 
 # === Fixtures for function `src.arguments.parse_arguments` ===
@@ -146,9 +148,10 @@ def cmd_all_ok_short_options(test_fasta_fpath, test_bam_fpath, test_outfpath) ->
         '-b', test_bam_fpath,
         '-o', test_outfpath,
         '-c', '10,20',
+        '-C', '1.5,2.0',
         '-n',
     ]
-# end def cmd_all_ok_short_options
+# end def
 
 @pytest.fixture
 def cmd_all_ok_long_options(test_fasta_fpath, test_bam_fpath, test_outfpath) -> Sequence[str]:
@@ -157,12 +160,13 @@ def cmd_all_ok_long_options(test_fasta_fpath, test_bam_fpath, test_outfpath) -> 
         '--target-fasta', test_fasta_fpath,
         '--bam', test_bam_fpath,
         '--outfile', test_outfpath,
-        '--coverage-thresholds', '10,20',
+        '--lower-coverage-thresholds', '10,20',
+        '--upper-coverage-coefficients', '1.5,2.0',
         '--no-zero-output',
         '--circular',
         '--organism', 'Czort lysy'
     ]
-# end def cmd_all_ok_long_options
+# end def
 
 @pytest.fixture
 def cmd_positional_args(test_fasta_fpath, test_bam_fpath) -> Sequence[str]:
@@ -174,7 +178,7 @@ def cmd_positional_args(test_fasta_fpath, test_bam_fpath) -> Sequence[str]:
         #          ~~~
         #          imitates space between threshold values
     ]
-# end def cmd_positional_args
+# end def
 
 @pytest.fixture
 def cmd_fasta_missing(test_bam_fpath) -> Sequence[str]:
@@ -182,7 +186,7 @@ def cmd_fasta_missing(test_bam_fpath) -> Sequence[str]:
         'consensus-highlighter',
         '-b', test_bam_fpath,
     ]
-# end def cmd_fasta_missing
+# end def
 
 @pytest.fixture
 def cmd_bam_missing(test_fasta_fpath) -> Sequence[str]:
@@ -190,7 +194,7 @@ def cmd_bam_missing(test_fasta_fpath) -> Sequence[str]:
         'consensus-highlighter',
         '-f', test_fasta_fpath,
     ]
-# end def cmd_bam_missing
+# end def
 
 
 def test_is_fasta(test_fasta_fpath, test_bam_fpath, test_coverage_fpath) -> None:
@@ -198,7 +202,7 @@ def test_is_fasta(test_fasta_fpath, test_bam_fpath, test_coverage_fpath) -> None
     assert args._is_fasta(test_fasta_fpath)
     assert not args._is_fasta(test_bam_fpath)
     assert not args._is_fasta(test_coverage_fpath)
-# end def test_is_fasta
+# end def
 
 
 def test_is_bam(test_fasta_fpath, test_bam_fpath, test_coverage_fpath) -> None:
@@ -206,7 +210,7 @@ def test_is_bam(test_fasta_fpath, test_bam_fpath, test_coverage_fpath) -> None:
     assert args._is_bam(test_bam_fpath)
     assert not args._is_bam(test_fasta_fpath)
     assert not args._is_bam(test_coverage_fpath)
-# end def test_is_bam
+# end def
 
 
 def test_coverage_not_parsable() -> None:
@@ -215,7 +219,7 @@ def test_coverage_not_parsable() -> None:
     assert args._coverage_not_parsable('0')
     assert not args._coverage_not_parsable('1')
     assert not args._coverage_not_parsable('12')
-# end def test_coverage_not_parsable
+# end def
 
 
 def test_add_zero_theshold(some_params) -> None:
@@ -223,8 +227,8 @@ def test_add_zero_theshold(some_params) -> None:
     args._add_zero_theshold(some_params)
 
     expected_zero_coverage: int = 0
-    assert some_params.coverage_thresholds[0].get_coverage() == expected_zero_coverage
-# end def test_add_zero_theshold
+    assert some_params.lower_coverage_thresholds[0] == expected_zero_coverage
+# end def
 
 
 class TestParseOptions:
@@ -243,14 +247,17 @@ class TestParseOptions:
         assert params.organism == '.'
 
         obtained_threshold_repr: str = ','.join(
-            map(
-                lambda covthr: str(covthr.get_coverage()),
-                params.coverage_thresholds
-            )
+            map(str, params.lower_coverage_thresholds)
         )
         expected_threshold_repr: str = opts_all_is_ok_short_options[3][1]
         assert obtained_threshold_repr == expected_threshold_repr
-    # end def test_opts_all_is_ok_short_options
+
+        obtained_coef_repr: str = ','.join(
+            map(str, params.upper_coverage_coefficients)
+        )
+        expected_coef_repr: str = opts_all_is_ok_short_options[4][1]
+        assert obtained_coef_repr == expected_coef_repr
+    # end def
 
 
     def test_opts_all_is_ok_long_options(self, opts_all_is_ok_long_options) -> None:
@@ -263,17 +270,20 @@ class TestParseOptions:
         assert params.outfpath == opts_all_is_ok_long_options[2][1]
         assert params.suppress_zero_cov_output == True
         assert params.topology == 'circular'
-        assert params.organism == opts_all_is_ok_long_options[6][1]
+        assert params.organism == opts_all_is_ok_long_options[7][1]
 
         obtained_threshold_repr: str = ','.join(
-            map(
-                lambda covthr: str(covthr.get_coverage()),
-                params.coverage_thresholds
-            )
+            map(str, params.lower_coverage_thresholds)
         )
         expected_threshold_repr: str = opts_all_is_ok_long_options[3][1]
         assert obtained_threshold_repr == expected_threshold_repr
-    # end def opts_all_is_ok_long_options
+
+        obtained_coef_repr: str = ','.join(
+            map(str, params.upper_coverage_coefficients)
+        )
+        expected_coef_repr: str = opts_all_is_ok_long_options[4][1]
+        assert obtained_coef_repr == expected_coef_repr
+    # end def
 
     def test_opts_invalid_fasta(self, opts_invalid_fasta) -> None:
         # Function tests how `_parse_options` parses options
@@ -282,7 +292,7 @@ class TestParseOptions:
         with pytest.raises(SystemExit):
             params: args.HighlighterParams = args._parse_options(opts_invalid_fasta)
         # end with
-    # end def test_opts_all_is_ok_short_options
+    # end def
 
     def test_opts_fasta_not_fasta(self, opts_fasta_not_fasta) -> None:
         # Function tests how `_parse_options` parses options
@@ -291,7 +301,7 @@ class TestParseOptions:
         with pytest.raises(SystemExit):
             params: args.HighlighterParams = args._parse_options(opts_fasta_not_fasta)
         # end with
-    # end def test_opts_all_is_ok_short_options
+    # end def
 
     def test_opts_invalid_bam(self, opts_invalid_bam) -> None:
         # Function tests how `_parse_options` parses options
@@ -300,7 +310,7 @@ class TestParseOptions:
         with pytest.raises(SystemExit):
             params: args.HighlighterParams = args._parse_options(opts_invalid_bam)
         # end with
-    # end def test_opts_invalid_bam
+    # end def
 
     def test_opts_bam_not_bam(self, opts_bam_not_bam) -> None:
         # Function tests how `_parse_options` parses options
@@ -309,7 +319,7 @@ class TestParseOptions:
         with pytest.raises(SystemExit):
             params: args.HighlighterParams = args._parse_options(opts_bam_not_bam)
         # end with
-    # end def test_opts_all_is_ok_short_options
+    # end def
 
     def test_opts_NAN_threshold(self, opts_NAN_threshold) -> None:
         # Function tests how `_parse_options` parses options
@@ -318,7 +328,7 @@ class TestParseOptions:
         with pytest.raises(SystemExit):
             params: args.HighlighterParams = args._parse_options(opts_NAN_threshold)
         # end with
-    # end def test_opts_NAN_threshold
+    # end def
 
     def test_opts_zero_threshold(self, opts_zero_threshold) -> None:
         # Function tests how `_parse_options` parses options
@@ -327,7 +337,7 @@ class TestParseOptions:
         with pytest.raises(SystemExit):
             params: args.HighlighterParams = args._parse_options(opts_zero_threshold)
         # end with
-    # end def test_opts_zero_threshold
+    # end def
 
     def test_opts_negative_threshold(self, opts_negative_threshold) -> None:
         # Function tests how `_parse_options` parses options
@@ -336,7 +346,7 @@ class TestParseOptions:
         with pytest.raises(SystemExit):
             params: args.HighlighterParams = args._parse_options(opts_negative_threshold)
         # end with
-    # end def test_opts_negative_threshold
+    # end def
 
     def test_opts_keep_zero_thresahold(self, opts_keep_zero_thresahold) -> None:
         # Function tests how `_parse_options` adds zero threshold
@@ -347,10 +357,10 @@ class TestParseOptions:
 
         zero_coverage: int = 0
         # Zero threshold should be first
-        first_coverage_value: int = params.coverage_thresholds[0].get_coverage()
+        first_coverage_value: int = params.lower_coverage_thresholds[0]
 
         assert zero_coverage == first_coverage_value
-    # end def test_opts_keep_zero_thresahold
+    # end def
 
     def test_opts_suppress_zero_thresahold(self, opts_suppress_zero_thresahold) -> None:
         # Function tests how `_parse_options` suppresses zero threshold
@@ -360,16 +370,9 @@ class TestParseOptions:
         assert params.suppress_zero_cov_output == True
 
         zero_coverage: int = 0
-        coverage_values: List[int] = list(
-            map(
-                lambda x: x.get_coverage(),
-                params.coverage_thresholds
-            )
-        )
-
-        assert not zero_coverage in coverage_values
-    # end def test_opts_keep_zero_thresahold
-# end class TestParseOptions
+        assert not zero_coverage in params.lower_coverage_thresholds
+    # end def
+# end class
 
 
 class TestParseArguments:
@@ -397,14 +400,17 @@ class TestParseArguments:
         assert params.organism == '.'
 
         obtained_threshold_repr: str = ','.join(
-            map(
-                lambda covthr: str(covthr.get_coverage()),
-                params.coverage_thresholds
-            )
+            map(str, params.lower_coverage_thresholds)
         )
         expected_threshold_repr: str = cmd_all_ok_short_options[8]
         assert obtained_threshold_repr == expected_threshold_repr
-    # end def test_cmd_all_ok_short_options
+
+        obtained_coef_repr: str = ','.join(
+            map(str, params.upper_coverage_coefficients)
+        )
+        expected_coef_repr: str = cmd_all_ok_short_options[10]
+        assert obtained_coef_repr == expected_coef_repr
+    # end def
 
     def test_cmd_all_ok_long_options(self, cmd_all_ok_long_options) -> None:
         # Function tests how `parse_arguments` parses correct long options
@@ -425,17 +431,20 @@ class TestParseArguments:
         assert params.outfpath == cmd_all_ok_long_options[6]
         assert params.suppress_zero_cov_output == True
         assert params.topology == 'circular'
-        assert params.organism == cmd_all_ok_long_options[12]
+        assert params.organism == cmd_all_ok_long_options[14]
 
         obtained_threshold_repr: str = ','.join(
-            map(
-                lambda covthr: str(covthr.get_coverage()),
-                params.coverage_thresholds
-            )
+            map(str, params.lower_coverage_thresholds)
         )
         expected_threshold_repr: str = cmd_all_ok_long_options[8]
         assert obtained_threshold_repr == expected_threshold_repr
-    # end def test_cmd_all_ok_long_options
+
+        obtained_coef_repr: str = ','.join(
+            map(str, params.upper_coverage_coefficients)
+        )
+        expected_coef_repr: str = cmd_all_ok_long_options[10]
+        assert obtained_coef_repr == expected_coef_repr
+    # end def
 
     def test_cmd_positional_args(self, cmd_positional_args) -> None:
         # Function tests how `parse_arguments` parses command line
@@ -453,7 +462,7 @@ class TestParseArguments:
 
         # Restore argv
         argv = buff_argv
-    # end def test_cmd_positional_args
+    # end def
 
     def test_cmd_fasta_missing(self, cmd_fasta_missing) -> None:
         # Function tests how `parse_arguments` parses command line
@@ -471,7 +480,7 @@ class TestParseArguments:
 
         # Restore argv
         argv = buff_argv
-    # end def test_cmd_fasta_missing
+    # end def
 
     def test_cmd_bam_missing(self, cmd_bam_missing) -> None:
         # Function tests how `parse_arguments` parses command line
@@ -489,5 +498,5 @@ class TestParseArguments:
 
         # Restore argv
         argv = buff_argv
-    # end def test_cmd_bam_missing
-# end class TestParseArgumsents
+    # end def
+# end class