+++ tests/test_obtain_coverage.py
@@ -24,8 +24,8 @@ class TestConfSamtoolsDepthCmd:
         expected: str = f'samtools depth -aa -J -o {test_coverage_fpath} {bam_fpath}'
 
         assert oc._conf_samtools_depth_cmd(bam_fpath, test_coverage_fpath) == expected
-    # end def test_samtools_depth_cmd
-# end class TestConfSamtoolsDepthCmd
+    # end def
+# end class
 
 
 class TestCountCovForAllRefs:
@@ -63,14 +63,14 @@ class TestCountCovForAllRefs:
         cov_line_pattern: str = '.+\t[0-9]+\t[0-9]+'
         def conforms_cov_line_pattern(string: str) -> bool:
             return not re.match(cov_line_pattern, string) is None
-        # end def conforms_cov_line_pattern
+        # end def
 
         assert all(map(conforms_cov_line_pattern, cov_lines))
 
         # Check if all positions of the first test sequence are present in the coverage file
         def contains_first_id(string: str) -> bool:
             return string.split('\t')[0] == first_test_seq_id
-        # end def contains_first_id
+        # end def
 
         expected_len_first_seq: int = first_test_seq_len
         assert len(tuple(filter(contains_first_id, cov_lines))) == expected_len_first_seq
@@ -78,13 +78,13 @@ class TestCountCovForAllRefs:
         # Check if all positions of the second test sequence are present in the coverage file
         def contains_second_id(string: str) -> bool:
             return string.split('\t')[0] == second_test_seq_id
-        # end def contains_second_id
+        # end def
 
         expected_len_second_seq: int = second_test_seq_len
         assert len(tuple(filter(contains_second_id, cov_lines))) == expected_len_second_seq
 
-    # end def test_coverage_counting
-# end class TestCountCovForAllRefs
+    # end def
+# end class
 
 
 class TestGetCoverageForReference:
@@ -105,7 +105,7 @@ class TestGetCoverageForReference:
         )
 
         assert len(cov_array) == first_test_seq_len
-    # end def test_get_covs_seq_1
+    # end def
 
     def test_get_covs_seq_2(
         self,
@@ -122,5 +122,5 @@ class TestGetCoverageForReference:
         )
 
         assert len(cov_array) == second_test_seq_len
-    # end def test_get_covs_seq_2
-# end class TestGetCoverageForReference
+    # end def
+# end class